add image when there is no image

+++ frontend/src/components/Review/index.js
@@ -9,8 +9,6 @@ import { deleteReview } from '../../store/comment';
 
 export const Review = (id) => {
     const history = useHistory();
-    const [review, setReview] = useState("");
-    const [stars, setStars] = useState("");
     const dispatch = useDispatch();
     const spotReviewsObj = useSelector(state => state.review)
     const spotReviews = Object.values(spotReviewsObj);
@@ -18,7 +16,20 @@ export const Review = (id) => {
     const spotReview = spotReviews.filter(spot=> (spot.spotId === parseInt(currSpot)))
     const [update, setUpdate] = useState(false);
     const [remove, setRemove] = useState(false);
+    const currentUserId = useSelector(state => state.session.user.id)
 
+    const [review, setReview] = useState("");
+    const [stars, setStars] = useState("");
+    const [validationErrors, setValidationErrors] = useState([]);
+    const [hasSubmitted, setHasSubmitted] = useState(false);
+
+    useEffect(() => {
+        const errors = [];
+        if (!review.length) errors.push('Please enter your review');
+        if (!stars.length) errors.push('Please enter your stars');
+        if (stars < 0 || stars > 6) errors.push('Please enter between 0~5')
+        setValidationErrors(errors);
+    }, [review,stars,])
 
     // ============ create new review======================
       const handleSubmit =  (e) => {
@@ -74,6 +85,7 @@ export const Review = (id) => {
                     Review
                     <input
                         type="text"
+                        placeholder='please leave a review'
                         value={review}
                         onChange={e => setReview(e.target.value)}
                     />
@@ -82,6 +94,8 @@ export const Review = (id) => {
                     stars
                     <input
                         type="number"
+                        min="0" max="5"
+                        placeholder='0'
                         value={stars}
                         onChange={e => setStars(e.target.value)}
                     />
@@ -106,9 +120,12 @@ export const Review = (id) => {
                         </li>
                        
                         <button value={id} onClick={()=>{
-                            dispatch(deleteReview(parseInt(id)));
-                            setRemove(true)
-                            history.push(`/spots/${currSpot}`);
+                            const deletePermission = userId !== currentUserId ? alert("No Permission to delete") : true
+                            if(deletePermission) {
+                                dispatch(deleteReview(parseInt(id)));
+                                setRemove(true)
+                                history.push(`/spots/${currSpot}`);
+                            }
                         }} className="delete">delete</button>  
                     </ul>
                 ))}
